Start component refactoring. Move to separate Namespace. Divide structure for Load, Save.
