+++ src/components/start/Start.css
@@ -6,7 +6,7 @@
   margin: 5em;
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: flex-start;
   align-items: flex-start;
   color: #5a6174;
 }
