+++ src/components/start/Start.js
@@ -1,36 +1,11 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 
-import { setWord } from '../actions';
-import Button from './button/Button';
-import ValidationError from './ValidationError';
+import NewGame from '../newgame/NewGame';
+import LoadGame from '../loadgame/LoadGame';
 import './Start.css';
 
 class Start extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = { inputValue: '' };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(event) {
-    this.setState({ inputValue: event.target.value.toUpperCase() });
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    this.props.setWord(this.state.inputValue || 'WORD');
-  }
-
   render() {
-    const error = this.props.error ? (
-      <ValidationError error={this.props.error} />
-    ) : (
-      undefined
-    );
     return (
       <section className="start">
         <section className="info">
@@ -39,34 +14,13 @@ class Start extends Component {
           </h1>
           <h2>train your brain</h2>
         </section>
-        <section className="action">
-          <input
-            className="wgInput"
-            type="text"
-            value={this.state.inputValue}
-            onChange={this.handleChange}
-            placeholder="WORD"
-          />
-          {error}
-          <Button
-            title="start game"
-            onClick={event => this.handleSubmit(event)}
-          />
+        <section className="info">
+          <NewGame />
+          <LoadGame />
         </section>
       </section>
     );
   }
 }
 
-const mapStateToProps = state => {
-  return {
-    error: state.game.error
-  };
-};
-const mapDispatchToProps = dispatch => {
-  return {
-    setWord: word => dispatch(setWord(word))
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Start);
+export default Start;
